[WIP] Refactor OAuth2 authentication plugin, endpoint
Until now support of OAuth2 in cauth was strongly tied with and thus
strictly limited to github as a provider. This refactor will allow
greater flexibility and make it easier to write OAuth2 based auth
plugins for new providers.
In order to support a OAuth2 provider, a specific plugin is still needed
though, because cauth needs to know how to fetch user data (login,
email, etc) from the provider; this is API-specific and thus cannot
be easily generalized.

Change-Id: Ic9077bcb58a893df22dcf4615368c680e808542a
Related-To: #1450